+++ src/main/kotlin/com/asap/asapbackend/global/config/ConfigurationPropertiesConfig.kt
@@ -1,6 +1,6 @@
 package com.asap.asapbackend.global.config
 
-import com.asap.asapbackend.global.security.jwt.JwtProperties
+import com.asap.asapbackend.global.jwt.JwtProperties
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Configuration
