+++ src/main/kotlin/com/asap/asapbackend/global/jwt/exception/ExpiredTokenException.kt
@@ -1,7 +1,7 @@
-package com.asap.asapbackend.global.security.jwt.exception
+package com.asap.asapbackend.global.jwt.exception
 
 import com.asap.asapbackend.global.exception.ErrorCode
 
-class InvalidTokenException(
+class ExpiredTokenException(
     override val errorCode: ErrorCode
 ): TokenException(errorCode)
