+++ src/main/kotlin/com/asap/asapbackend/global/jwt/exception/TokenException.kt
@@ -1,4 +1,4 @@
-package com.asap.asapbackend.global.security.jwt.exception
+package com.asap.asapbackend.global.jwt.exception
 
 import com.asap.asapbackend.global.exception.BusinessException
 import com.asap.asapbackend.global.exception.ErrorCode
